Add LocalizedDescription to LogError
